requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
arrow==1.2.2
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
asynco==0.0.5
attrs==21.4.0
Babel==2.9.1
bcrypt==3.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
bs4==0.0.1
captcha==0.3
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
click==8.0.3
colorama==0.4.4
colorlog==6.6.0
command-not-found==0.3
configparser==5.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
DateTime==4.5
DAWG-Python==0.7.2
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==3.2.13
docopt==0.6.2
elementpath==3.0.2
fasteners==0.14.1
filelock==3.8.2
Flask==2.1.3
fonttools==4.34.4
frozenlist==1.3.0
future==0.18.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==1.1.1
isodate==0.6.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonpath-rw==1.4.0
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.1
matplotlib==3.5.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.2
mutagen==1.45.1
natsort==8.0.2
netifaces==0.11.0
numpy==1.23.1
oauthlib==3.2.0
olefile==0.46
opentele==1.15.1
packaging==21.3
paramiko==2.9.3
peewee==3.15.2
pexpect==4.8.0
phonenumbers==8.12.40
pi==0.1.2
Pillow==9.0.1
platformdirs==2.6.0
pluggy==1.0.0
ply==3.11
protobuf==3.12.4
protonup==0.1.4
ptyprocess==0.7.0
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
pygame==2.1.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymemcache==3.5.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
PyNaCl==1.5.0
pyparsing==3.0.0
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyRFC3339==1.1
Pyrogram==1.4.16
PySocks==1.7.1
pyTelegramBotAPI==4.6.0
pytest==7.1.2
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.10.1
rsa==4.9
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
SimpleQIWI==0.7
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
ssh-import-id==5.11
systemd-python==234
Telethon==1.25.0
TgCrypto==1.2.3
threaded==4.1.0
times==0.7
tkintertable==1.3.3
tomli==2.0.1
types-aiofiles==0.8.10
types-all==1.0.0
types-annoy==1.17.8
types-atomicwrites==1.4.5
types-backports==0.1.3
types-backports-abc==0.5.2
types-bleach==5.0.3
types-boto==2.49.17
types-cachetools==5.2.1
types-certifi==2021.10.8.3
types-characteristic==14.3.7
types-chardet==5.0.3
types-click==7.1.8
types-click-spinner==0.1.13
types-colorama==0.4.15
types-contextvars==2.4.7
types-croniter==1.3.1
types-cryptography==3.3.21
types-dataclasses==0.6.6
types-dateparser==1.1.4
types-DateTimeRange==1.2.7
types-decorator==5.1.8
types-Deprecated==1.2.9
types-docopt==0.6.11
types-docutils==0.19.0
types-emoji==1.7.4
types-enum34==1.1.8
types-fb303==1.0.0
types-filelock==3.2.7
types-first==2.0.5
types-Flask==1.1.6
types-freezegun==1.1.10
types-frozendict==2.0.9
types-futures==3.3.8
types-geoip2==3.0.0
types-ipaddress==1.0.8
types-itsdangerous==1.1.6
types-JACK-Client==0.5.10
types-Jinja2==2.11.9
types-kazoo==0.1.3
types-Markdown==3.4.0
types-MarkupSafe==1.1.10
types-maxminddb==1.5.0
types-mock==4.0.15
types-mypy-extensions==0.4.21
types-nmap==0.1.6
types-openssl-python==0.1.3
types-orjson==3.6.2
types-paramiko==2.11.3
types-pathlib2==2.3.0
types-Pillow==9.2.0
types-pkg-resources==0.1.3
types-polib==1.1.12
types-protobuf==3.19.22
types-pyaudio==0.2.16
types-pycurl==7.45.1
types-pyfarmhash==0.3.1
types-PyJWT==1.7.1
types-pymssql==2.1.0
types-PyMySQL==1.0.19
types-pyRFC3339==1.1.1
types-pysftp==0.2.17
types-python-dateutil==2.8.19
types-python-gflags==3.1.7
types-python-slugify==6.1.0
types-pytz==2022.1.2
types-pyvmomi==7.0.7
types-PyYAML==6.0.11
types-redis==4.3.11
types-requests==2.28.5
types-retry==0.9.9
types-Routes==2.5.0
types-scribe==2.0.0
types-simplejson==3.17.7
types-singledispatch==3.7.5
types-six==1.16.18
types-tabulate==0.8.11
types-termcolor==1.1.5
types-toml==0.10.8
types-tornado==5.1.1
types-typed-ast==1.5.7
types-tzlocal==4.2.2
types-ujson==5.4.0
types-urllib3==1.26.17
types-waitress==2.1.4
types-Werkzeug==1.0.9
types-xxhash==3.0.4
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uuid==1.30
vboxapi==1.0
virtualenv==20.17.1
wadllib==1.3.6
webencodings==0.5.1
websockets==9.1
Werkzeug==2.2.0
wxPython==4.0.7
xdg==5
xkit==0.0.0
xmlschema==2.1.1
xmltodict==0.13.0
yarl==1.7.2
yt-dlp==2022.4.8
zeep==4.2.1
zipp==1.0.0
zope.interface==5.4.0

pandas~=1.5.2

apscheduler~=3.9.1.post1
celery~=5.2.7
argparse~=1.4.0